# agent-dashboard/data/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Dashboard API
  version: 1.0.0
  description: |
    독립 웹뷰/대시보드를 위한 상태 스냅샷 + 이벤트 스트림 API.
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/ingest:
    post:
      summary: generic 소스로 이벤트를 주입
      operationId: ingestEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/IngestEventInput'
                - type: array
                  items:
                    $ref: '#/components/schemas/IngestEventInput'
      responses:
        '200':
          description: 수락 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer
                  last_seq:
                    type: integer
                    nullable: true
                  run_id:
                    type: string
                    nullable: true
                required: [accepted, last_seq, run_id]
components:
  schemas:
    IngestEventInput:
      type: object
      properties:
        ts:
          type: string
          format: date-time
        run_id:
          type: string
        type:
          type: string
          enum:
            - agent.created
            - agent.status.changed
            - agent.tool.started
            - agent.tool.finished
            - agent.permission.waiting
            - agent.permission.cleared
            - agent.position.changed
            - agent.seat.assigned
            - agent.subagent.created
            - agent.subagent.removed
            - agent.tools.cleared
            - agent.removed
            - layout.updated
        agent_id:
          type: string
        payload:
          type: object
      required: [type, agent_id, payload]